feat: update doc for link & comment
