+++ packages/sheets-hyper-link-ui/README.md
@@ -30,6 +30,29 @@ import { UniverSheetsHyperLinkUIPlugin } from '@univerjs/sheets-hyper-link-ui';
 univer.registerPlugin(UniverSheetsHyperLinkUIPlugin);
 ```
 
+### API
+```typescript
+// All commands
+import { AddHyperLinkCommand, type IAddHyperLinkCommandParams } from '@univerjs/sheets-hyper-link-ui';
+import { RemoveHyperLinkCommand, CancelHyperLinkCommand, type IRemoveHyperLinkCommandParams } from '@univerjs/sheets-hyper-link-ui';
+import { UpdateHyperLinkCommand, type IUpdateHyperLinkCommandParams } from '@univerjs/sheets-hyper-link-ui';
+
+// Add hyperlink by command
+const commandService = univer.__getInjector().get(ICommandService);
+
+commandService.executeCommand(AddHyperLinkCommand.id, {
+    unitId: 'unitId',
+    subUnitId: 'subUnitId',
+    link: {
+        // comment content
+        payload: 'https://univer.ai',
+        id: '1',
+        row: 0,
+        column: 0
+    },
+} as IAddCommentCommandParams);
+```
+
 <!-- Links -->
 [npm-version-shield]: https://img.shields.io/npm/v/@univerjs/sheets-hyper-link-ui?style=flat-square
 [npm-version-link]: https://npmjs.com/package/@univerjs/sheets-hyper-link-ui
