+++ packages/sheets-hyper-link-ui/src/controllers/ui.controller.ts
@@ -22,6 +22,9 @@ import { CellLinkPopup } from '../views/CellLinkPopup';
 import { CellLinkEdit } from '../views/CellLinkEdit';
 import { CloseHyperLinkSidebarOperation, InsertHyperLinkOperation, InsertHyperLinkToolbarOperation, OpenHyperLinkSidebarOperation } from '../commands/operations/sidebar.operations';
 import type { IUniverSheetsHyperLinkUIConfig } from '../types/interfaces/i-config';
+import { AddHyperLinkCommand } from '../commands/commands/add-hyper-link.command';
+import { UpdateHyperLinkCommand } from '../commands/commands/update-hyper-link.command';
+import { CancelHyperLinkCommand, RemoveHyperLinkCommand } from '../commands/commands/remove-hyper-link.command';
 import { insertLinkMenuFactory, insertLinkMenuToolbarFactory, InsertLinkShortcut } from './menu';
 
 @OnLifecycle(LifecycleStages.Ready, SheetsHyperLinkUIController)
@@ -59,6 +62,11 @@ export class SheetsHyperLinkUIController extends Disposable {
             CloseHyperLinkSidebarOperation,
             InsertHyperLinkOperation,
             InsertHyperLinkToolbarOperation,
+
+            AddHyperLinkCommand,
+            UpdateHyperLinkCommand,
+            RemoveHyperLinkCommand,
+            CancelHyperLinkCommand,
         ].forEach((command) => {
             this._commandService.registerCommand(command);
         });
