+++ packages/sheets-hyper-link/src/controllers/sheet-hyper-link.controller.ts
@@ -18,9 +18,6 @@ import { Disposable, ICommandService, LifecycleStages, OnLifecycle } from '@univ
 import { AddHyperLinkMutation } from '../commands/mutations/add-hyper-link.mutation';
 import { UpdateHyperLinkMutation, UpdateHyperLinkRefMutation } from '../commands/mutations/update-hyper-link.mutation';
 import { RemoveHyperLinkMutation } from '../commands/mutations/remove-hyper-link.mutation';
-import { AddHyperLinkCommand } from '../commands/commands/add-hyper-link.command';
-import { UpdateHyperLinkCommand } from '../commands/commands/update-hyper-link.command';
-import { CancelHyperLinkCommand, RemoveHyperLinkCommand } from '../commands/commands/remove-hyper-link.command';
 
 @OnLifecycle(LifecycleStages.Starting, SheetsHyperLinkController)
 export class SheetsHyperLinkController extends Disposable {
@@ -38,11 +35,6 @@ export class SheetsHyperLinkController extends Disposable {
             UpdateHyperLinkMutation,
             RemoveHyperLinkMutation,
             UpdateHyperLinkRefMutation,
-
-            AddHyperLinkCommand,
-            UpdateHyperLinkCommand,
-            RemoveHyperLinkCommand,
-            CancelHyperLinkCommand,
         ].forEach((command) => {
             this._commandService.registerCommand(command);
         });
